requirements.txt
pyserial2
pynmeagps